% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-tools.R
\name{basic_umap}
\alias{basic_umap}
\title{Basic Umap}
\usage{
basic_umap(seurat, colors = NULL)
}
\arguments{
\item{seurat}{seurat object}

\item{colors}{}
}
\value{

}
\description{
Basic Umap
}
\examples{
basic_umap(seurat, colors)
}
