% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{cowtime}
\alias{cowtime}
\title{Cowtime}
\usage{
cowtime(expr, name = NA)
}
\arguments{
\item{expr}{Expression to be timed.}

\item{name}{(Optional) Name of expression to be printed in console}
}
\value{

}
\description{
Cowtime
}
\examples{
cowtime(name = 'do thing', { do_thing() })
}
