% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{display_banner}
\alias{display_banner}
\title{Display Banner}
\usage{
display_banner(banner, color_map, color_legend)
}
\arguments{
\item{banner}{list of character-color pairs}

\item{color_map}{color codes for banner}
}
\value{

}
\description{
Display Banner
}
\examples{
display_banner()
}
