% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcsprocessor.R
\name{fcsprocessor}
\alias{fcsprocessor}
\title{Prepare Flow Cytometry Standard (FCS) formatted data for gateR package}
\usage{
fcsprocessor(path, mdata, bl)
}
\arguments{
\item{path}{Character string of the path to the directory containing the FCS files and a metadata file. Note: The string must end with '/'}

\item{mdata}{Character string of the name of the metadata file in the directory. Note: The file must be in a CSV format}

\item{bl}{Object of class 'list'. This is a named list with column names of FCS files to exclude from the analysis}
}
\value{
An object of class 'data.frame'. This is a named data frame with columns (features) for each condition specified in the metadata file and an ID for each row.
}
\description{
Converts a single or a collection of objects of class 'fcs' to an object of class 'data.frame' readable by \code{\link{gating}}, \code{\link{rrs}}, and \code{\link{lotrrs}} functions.
}
\examples{
if (interactive()) {
# INSERT TEST HERE
}

}
