% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{filter_degs}
\alias{filter_degs}
\title{Filter DEGs}
\usage{
filter_degs(
  degs,
  bh_threshold = Inf,
  logfc_threshold = 0,
  excluded_patterns = NULL
)
}
\arguments{
\item{degs}{Data table containing DEGs.  Expecting format like the output of read_degs()}

\item{bh_threshold}{maximum BH corrected p value}

\item{logfc_threshold}{Minimum log fold change value}

\item{excluded_patterns}{List of regex patterns to exlude genes}
}
\value{

}
\description{
Filter DEGs
}
\examples{
'path/to/file' |> read_degs() |> filter_degs(logfc_threshold = 0.25, excluded_patterns = c("^RP", "^XIST$"))
}
