% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{get_filename}
\alias{get_filename}
\title{get_filename Given a full path, this function returns just the filename.}
\usage{
get_filename(path)
}
\arguments{
\item{path}{}
}
\description{
get_filename Given a full path, this function returns just the filename.
}
\examples{
get_filename('/path/to/file.txt')
}
