% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{\%!in\%}
\alias{\%!in\%}
\title{\%!in\%}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Item to compare}

\item{y}{Item to compare}
}
\value{

}
\description{
\%!in\%
}
\examples{
if(1 \%!in! c(2, 3, 4)) {do_something()}
}
