% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{merge_csvs}
\alias{merge_csvs}
\title{Merge CSVs}
\usage{
merge_csvs(files, ids = files)
}
\arguments{
\item{files}{List of tsv/csv files to be read as dataframes and combined.}

\item{ids}{List of ids in same length/order as files. If not provided, defaults to files param value.}
}
\value{

}
\description{
Merge CSVs
}
\examples{
path/to/files |> list.files() |> merge_csvs()
}
