% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{merge_dfs}
\alias{merge_dfs}
\title{Merge DFs}
\usage{
merge_dfs(files, ids)
}
\arguments{
\item{files}{List of tsv/csv files containing dataframes}

\item{ids}{List of ids in same length/order as files}
}
\value{

}
\description{
Merge DFs
}
\examples{
merge_dfs(files, ids)
}
