% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{merge_pdfs}
\alias{merge_pdfs}
\title{Merge PDFs}
\usage{
merge_pdfs(dir)
}
\arguments{
\item{dir}{Directory containing PDFs to combine}
}
\value{

}
\description{
Merge PDFs
}
\examples{
merge_pdfs('path/to/folder/containing/PDFs')
}
