% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{mkdirs}
\alias{mkdirs}
\title{mkdirs}
\usage{
mkdirs(path, glue = TRUE)
}
\arguments{
\item{path}{path to directory}

\item{glue}{= TRUE.  If TRUE, input passed through the glue() function}
}
\description{
mkdirs
}
\examples{
mkdirs('path/to/dir')
}
