% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{parse_deg_filename}
\alias{parse_deg_filename}
\title{Parse DEG Filename}
\usage{
parse_deg_filename(deg_file)
}
\arguments{
\item{deg_file}{csv file containing DEGs. Either full path or filename is acceptable}
}
\value{

}
\description{
Parse DEG Filename
}
\examples{
parse_deg_filename("path/to/file.csv")
}
