% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_plot.R
\name{pval_plot}
\alias{pval_plot}
\title{Prepare significant p-values for plotting}
\usage{
pval_plot(input, alpha)
}
\arguments{
\item{input}{An object of class 'rrs' from the \code{\link{lrren}} function.}

\item{alpha}{Numeric. The two-tailed alpha level for significance threshold (default in \code{\link{rrs}}, \code{\link{lotrrs}}, and \code{\link{gate}} functions is 0.05).}
}
\value{
An object of class 'raster' with categorical values:

\itemize{
\item A value of 1: Significant numerator.
\item A value of 2: Insignificant.
\item A value of 3: Significant demoninator.
}
}
\description{
Internal function to convert an object of class 'im' to values readable by \code{\link[fields]{image.plot}} function within the \code{\link{rrs}}, \code{\link{lotrrs}}, and \code{\link{gate}} functions.
}
\keyword{internal}
