% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{read_degs_file}
\alias{read_degs_file}
\title{Read DEGs file}
\usage{
read_degs_file(file)
}
\arguments{
\item{file}{csv file containing degs}
}
\value{

}
\description{
Read DEGs file
}
\examples{
read_degs_file('path/to/file.csv')
}
