% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degs.R
\name{read_degs_folder}
\alias{read_degs_folder}
\title{Read DEGs folder}
\usage{
read_degs_folder(degs_folder)
}
\arguments{
\item{degs_folder}{path to folder containing degs.csv and metadata.yml}
}
\value{

}
\description{
Read DEGs folder
}
\examples{
read_degs_folder('path/to/folder')
}
