% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{merge_dfs}
\alias{merge_dfs}
\title{Merge DFs}
\usage{
merge_dfs(files, ids = files)
}
\arguments{
\item{files}{List of tsv/csv files to be read as dataframes and combined.}

\item{ids}{List of ids in same length/order as files. If not provided, defaults to files param value.}
}
\description{
Merge DFs
}
\examples{
path/to/files |> list.files() |> merge_dfs()
}
